+++ client/src/AdminDashboard.js
@@ -1,0 +1,10 @@
+import { OfficialOrgs, OfficialEvents } from './OfficialEntities.js';
+
+/**
+ * AdminDashboard class handles admin dashboard functionalities.
+ */
+export class AdminDashboard {
+  displayPendingSubmissions() { return []; }
+  reviewSubmission(itemID) { return {}; }
+  processDecision(item, decision) { }
+}
