+++ client/src/Event.js
@@ -1,0 +1,21 @@
+/**
+ * Event class represents an event in the system.
+ */
+export class Event {
+  constructor({ eventID, organizationID, title, description, thumbnail, banner,
+    startDate, endDate, privacy, officialStatus, landmarkID }) {
+    this.eventID = eventID;
+    this.organizationID = organizationID;
+    this.title = title;
+    this.description = description;
+    this.thumbnail = thumbnail;
+    this.banner = banner;
+    this.startDate = startDate;
+    this.endDate = endDate;
+    this.privacy = privacy;
+    this.officialStatus = officialStatus;
+    this.landmarkID = landmarkID;
+  }
+  async createEvent() { /* see EventController.registerEvent */ }
+  async rsvpToEvent(status) { /* implement as needed */ }
+}
