+++ client/src/OfficialEntities.js
@@ -1,0 +1,17 @@
+/**
+ * OfficialOrgs class represents official organizations in the system.
+ */
+export class OfficialOrgs {
+  constructor({ orgID }) {
+    this.orgID = orgID;
+  }
+}
+
+/**
+ * OfficialEvents class represents official events in the system.
+ */
+export class OfficialEvents {
+  constructor({ eventID }) {
+    this.eventID = eventID;
+  }
+}
