+++ client/src/OfficialStatusController.js
@@ -1,0 +1,24 @@
+import { OfficialOrgs, OfficialEvents } from './OfficialEntities.js';
+
+/**
+ * OfficialStatusController class handles official status verification.
+ */
+export class OfficialStatusController {
+  verifyOrganizationOfficialStatus(org) { 
+    // Implementation to verify official status for an organization
+    return true;
+  }
+  verifyEventOfficialStatus(event) { 
+    // Implementation to verify official status for an event 
+    return true;
+  }
+  acceptOrganization(org) { 
+    // Process acceptance (stub)
+  }
+  acceptEvent(event) { 
+    // Process acceptance (stub)
+  }
+  denySubmission(item) { 
+    // Process denial (stub)
+  }
+}
